Added packages for each feature
