+++ app/src/main/java/ru/practicum/android/diploma/favorite/presentation/fragment/Favourite.kt
@@ -1,21 +1,21 @@
-package ru.practicum.android.diploma
+package ru.practicum.android.diploma.favorite.presentation.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import ru.practicum.android.diploma.databinding.FragmentCrewBinding
+import ru.practicum.android.diploma.databinding.FragmentFavouriteBinding
 
-class Crew : Fragment() {
-    private  var _binding: FragmentCrewBinding? = null
+class Favourite : Fragment() {
+    private  var _binding: FragmentFavouriteBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentCrewBinding.inflate(inflater,container,false)
+        _binding = FragmentFavouriteBinding.inflate(inflater,container,false)
 
         // Inflate the layout for this fragment
         return binding.root
