+++ app/src/main/java/ru/practicum/android/diploma/filter/presentation/fragment/District.kt
@@ -1,17 +1,16 @@
-package ru.practicum.android.diploma
+package ru.practicum.android.diploma.filter.presentation.fragment
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import ru.practicum.android.diploma.databinding.FragmentCountryBinding
-import ru.practicum.android.diploma.databinding.FragmentWorkPlaceBinding
+import ru.practicum.android.diploma.databinding.FragmentDistrictBinding
 
-class Country : Fragment() {
-    private var _binding: FragmentCountryBinding? = null
-    private val binding:FragmentCountryBinding get() = _binding!!
+class District : Fragment() {
+    private var _binding: FragmentDistrictBinding? = null
+    private val binding:FragmentDistrictBinding get() = _binding!!
 
     private fun setUiListeners(){
         with(binding){
@@ -24,8 +23,8 @@ class Country : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentCountryBinding.inflate(inflater,container,false)
+    ): View {
+        _binding = FragmentDistrictBinding.inflate(inflater,container,false)
 
         // Inflate the layout for this fragment
         return binding.root
