+++ app/src/main/java/ru/practicum/android/diploma/root/presentation/RootActivity.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.ui.root
+package ru.practicum.android.diploma.root.presentation
 
 import android.os.Bundle
 import android.view.View
