+++ app/src/main/res/layout/activity_root.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.root.RootActivity">
+    tools:context=".root.presentation.RootActivity">
 
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
