+++ app/src/main/res/layout/fragment_country.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WorkPlace">
+    tools:context=".filter.presentation.fragment.WorkPlace">
 
 
         <com.google.android.material.appbar.MaterialToolbar
