+++ app/src/main/res/navigation/job_pulse_nav.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/search"
-        android:name="ru.practicum.android.diploma.Search"
+        android:name="ru.practicum.android.diploma.search.presentation.fragment.Search"
         android:label="fragment_search"
         tools:layout="@layout/fragment_search" >
         <action
@@ -17,17 +17,17 @@
 
     <fragment
         android:id="@+id/favourite"
-        android:name="ru.practicum.android.diploma.Favourite"
+        android:name="ru.practicum.android.diploma.favorite.presentation.fragment.Favourite"
         android:label="FavouriteFragment"
         tools:layout="@layout/fragment_favourite"/>
     <fragment
         android:id="@+id/crew"
-        android:name="ru.practicum.android.diploma.Crew"
+        android:name="ru.practicum.android.diploma.team.presentation.Crew"
         android:label="CrewFragment"
         tools:layout="@layout/fragment_crew"/>
     <fragment
         android:id="@+id/filters"
-        android:name="ru.practicum.android.diploma.Filters"
+        android:name="ru.practicum.android.diploma.filter.presentation.fragment.Filters"
         android:label="Filters"
         tools:layout="@layout/fragment_filters">
         <action
@@ -37,7 +37,7 @@
 
     <fragment
         android:id="@+id/workPlace"
-        android:name="ru.practicum.android.diploma.WorkPlace"
+        android:name="ru.practicum.android.diploma.filter.presentation.fragment.WorkPlace"
         android:label="fragment_work_place"
         tools:layout="@layout/fragment_work_place" >
         <action
@@ -50,12 +50,12 @@
 
     <fragment
         android:id="@+id/country"
-        android:name="ru.practicum.android.diploma.Country"
+        android:name="ru.practicum.android.diploma.filter.presentation.fragment.Country"
         android:label="Country"
         tools:layout="@layout/fragment_country"/>
     <fragment
         android:id="@+id/district"
-        android:name="ru.practicum.android.diploma.District"
+        android:name="ru.practicum.android.diploma.filter.presentation.fragment.District"
         android:label="District"
         tools:layout="@layout/fragment_district"/>
 </navigation>
